Revisit the problem of building without X11
There had been some efforts to remove unnecessary dependencies against X11.
But several toolkits (especially TKV3d) are needed by OCAF and Data Exchange,
and they cannot easily be compiled without X11.

This commit uses a different approach, based on Debian packages.  A new
OCE_HEADLESS option is added; it is OFF by default, and can be set to ON
when compiling without X11.
The list of toolkits passed to the PROCESS_MODULE macro is splitted into
two parts, the first one contains toolkits which can be compiled without
X11, and the second one (optional) contains toolkits which are compiled
only if OCE_HEADLESS is OFF.

This allows building many toolkits from OCAF and Data Exchange without X11.
Of course, VISUALISATION and DRAW modules are disabled when OCE_HEADLESS
is ON.